backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
trafilatura
google-generativeai
python-dotenv
numpy
